+++ torchcp/classification/trainer/__init__.py
@@ -7,3 +7,6 @@
 
 from .base_trainer import Trainer
 from .ordinal_trainer import OrdinalTrainer
+from .ts_trainer import TSTrainer
+from .confts_trainer import ConfTSTrainer
+from .model import TemperatureScalingModel